Ai-vshow/requirements.txt
# requirements.txt
python-dotenv
appium-python-client
Pillow
langchain-core
langchain-openai
langgraph
pydantic-settings